% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatData}
\alias{formatData}
\title{Convert pre-processed data and capture efficiencies into the format required by the inference algorithm}
\usage{
formatData(preprocessedData, alphas, cores)
}
\description{
Convert pre-processed data and capture efficiencies into the format required by the inference algorithm
}
