% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPython.R
\name{getPython}
\alias{getPython}
\title{Install miniconda python environment and required modules if not already installed and then load python functions
#' @export}
\usage{
getPython()
}
\description{
Install miniconda python environment and required modules if not already installed and then load python functions
#' @export
}
