% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MALAwithinGibbs.R
\name{MALAwithinGibbs}
\alias{MALAwithinGibbs}
\title{Function for sampling from posterior distributions of bursting parameters over a set of genes in parallel}
\usage{
MALAwithinGibbs(model, cores, Time, Data)
}
\arguments{
\item{model}{Which model to use to carry out inference}

\item{cores}{The number of cores to parallelise over}

\item{Time}{The 4sU pulse duration in minutes}

\item{Data}{The dataset, in a format consistent with the output of the 'formatData' function}
}
\value{
The model used (tracked in case forced to switch from model 2 to 3) and the sample from the posterior distribution generated by the Markov chain
}
\description{
Function for sampling from posterior distributions of bursting parameters over a set of genes in parallel
}
