% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSimulator.R
\name{simulateGenes}
\alias{simulateGenes}
\title{Simulate data in the format required for the inference algorithm based on real data and 'ground truth' parameter values to test algorithm performance on}
\usage{
simulateGenes(cores, Time, Data, estimates)
}
\description{
Simulate data in the format required for the inference algorithm based on real data and 'ground truth' parameter values to test algorithm performance on
}
